initial changes for new mcp spec
